requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
django-filter
pillow
drf-yasg #Документашка сама ставит
djangorestframework-simplejwt # здесьнормальный токкен который не сохраняется в бд